rename output file so no overwrite with multiple test scripts; move test reports to specified output dir; add parser for mocha output post-processing (xml to json)
